+++ CHANGELOG.md
@@ -18,6 +18,22 @@ Categories, defined in [changemap.json](.github/clq/changemap.json):
   - `Fixed` for any bugfixes.
   - `Security` in case of vulnerabilities.
 
+## [2.0.0] - 2025-03-17
+
+### Changed
+
+- CLUSTER_IAM does not end with trailing `:`
+- The secret to access the GitHub OCI registry is now named `ghcr-secret`.
+
+### Added
+
+- Run in verbose/debug mode when the workflow runs in debug mode; this mode propagates to helm execution.
+- Logs the aws caller identity in verbose mode
+
+### Fixed
+
+- The `clean-up` option should leave what was installed; it now triggers `--wait` instead of `--atomic`
+
 ## [1.1.0] - 2025-01-17
 
 ### Added